Create a `Descent` object.
Create an object that can be passed through the gateways to transport.
It is called `Descent` to disambigulate the action of calling the next
gateway from the action of making an orthongal request as in an
authentication request when credentials are missing.

Will also need a way to replay a request should we implement retries or
redirects.