Errors are returned as streams.
This allows you to flow all responses through a converter, including
error messages generated by gateways. Thus, a gateway mechanism like
client credentials or redirect should come before your converter
strategies so that if they generate errors, those errors are converted.

Upon reflection, it would seem that the converter middleware should
attempt to use itself to convert any error responses that it created.
This probably cannot be implemented in the Parser root unless you want
to try to make a guess as to what is a generated error and what is an
actual error passing through the conversion, or else set additional
flags on the response.