Verbose parser selection.
Simplify parser selection by using objects instead of argument
interpretation magic to select a parser. Now specifying an array of
options that can select based on headers, mime-type and status code to
specify a parser.

Seemed to need to rework error handling a bit, and will probably rework
it more when I rework the transport.