Modified error handling.
Updating error handling so that there is only one way for gateways to
raise errors, through an error object, and any other error is going to
be considered a programming exception which should go uncaught.

Going to make it so that gateways can just propagate a raise and have
the `Descent` create a nested error, so that errors by gateways are
nested within the errors of the main descent.