Simplified transport.
Now waiting for the response and the request write to finish before
processing the response. This makes explicit the use suggested by
Vizsla, the use of HTTP as an RPC protocol where you call a function and
wait for a return instead of treating HTTP as a duplex stream.

The vast majority of my HTTP usage is a GET or a POST of JSON with a
JSON result, in both cases the doesn't stream input, they deal with
headers only or headers plus the entirety of the body. They can't send a
response until the entire message has arrived. The JSON result must be
parsed from the entirety of response. Waiting for the request write to
finish is not going to reduce the throughput of a program in this most
common case, but it will make it easier to reason about where errors are
coming from.

There are a few instances where a GET or a POST of JSON results in a
streaming response, like streaming JSON, a honest-to-goodness download,
or even long polling. For these we work with the response stream. Errors
are now reported through the response stream.

Finally, there may have been a time in the past decade that I've
actually uploaded a file, but I can't remember it, but in that case it
was an upload with a response that verified that the upload succeeded.

Never have I had to deal with a response stream that affected the
contents of the request stream and when I do, I'll be sure to write a
carefully constructed UI with the `http` module directly.

Also, created an extraneous error logger. Will generally route an error
through the request call or else the response stream, but once end is
reached there might still be an error that follows for whatever reason.
This is just the nature of these error handlers. I'll never feel like I
understand the order and quantity of error responses.