Sort out actual stream end and piping.
If you shutdown the pipe to the pass-through based on response end, then
you might truncate the stream. You need to respond to the end that comes
through the pipe. For now I'm going to see if we can do without hooking
the end event and can just shutdown based on close.